🔀 chore(nft-event-pass): rename eventPass library to event-pass for consistency and clarity
