+++ libs/nft/event-pass/README.md
@@ -1,0 +1,11 @@
+# nft-event-pass
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build nft-event-pass` to build the library.
+
+## Running unit tests
+
+Run `nx test nft-event-pass` to execute the unit tests via [Jest](https://jestjs.io).
