+++ libs/nft/event-pass/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@nft/event-pass",
+  "version": "0.0.1"
+}
