+++ libs/nft/event-pass/project.json
@@ -1,18 +1,18 @@
 {
-  "name": "nft-eventPass",
+  "name": "nft-event-pass",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/nft/eventPass/src",
+  "sourceRoot": "libs/nft/event-pass/src",
   "projectType": "library",
   "targets": {
     "build": {
       "executor": "@nx/js:swc",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/libs/nft/eventPass",
-        "tsConfig": "libs/nft/eventPass/tsconfig.lib.json",
-        "packageJson": "libs/nft/eventPass/package.json",
-        "main": "libs/nft/eventPass/src/index.ts",
-        "assets": ["libs/nft/eventPass/*.md"]
+        "outputPath": "dist/libs/nft/event-pass",
+        "tsConfig": "libs/nft/event-pass/tsconfig.lib.json",
+        "packageJson": "libs/nft/event-pass/package.json",
+        "main": "libs/nft/event-pass/src/index.ts",
+        "assets": ["libs/nft/event-pass/*.md"]
       }
     },
     "lint": {
@@ -20,8 +20,8 @@
       "outputs": ["{options.outputFile}"],
       "options": {
         "lintFilePatterns": [
-          "libs/nft/eventPass/**/*.ts",
-          "libs/nft/eventPass/package.json"
+          "libs/nft/event-pass/**/*.ts",
+          "libs/nft/event-pass/package.json"
         ]
       }
     },
@@ -29,7 +29,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/nft/eventPass/jest.config.ts",
+        "jestConfig": "libs/nft/event-pass/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
