+++ libs/nft/event-pass/tsconfig.json
