+++ libs/nft/event-pass/tsconfig.lib.json
