+++ libs/nft/event-pass/tsconfig.spec.json
