--- libs/nft/eventPass/README.md
@@ -1,11 +1,0 @@
-# nft-eventPass
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build nft-eventPass` to build the library.
-
-## Running unit tests
-
-Run `nx test nft-eventPass` to execute the unit tests via [Jest](https://jestjs.io).
