--- libs/nft/eventPass/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@nft/eventPass",
-  "version": "0.0.1"
-}
